give conversion methods default implementation in GameComponent
Some very old compiled game component types are not derived from `AbstractGameComponent` and therefore fail to  instantiate without the new conversion-related methods. Instead of requiring a plug-in update (which would be tricky since old versions without the conversion code could not use the new plug-in and vice-versa), these methods are given a default implementation in the `GameComponent` interface.